+++ src/picker/js/storage.js
@@ -1,53 +1,58 @@
-/* eslint-disable */
+import logger from 'loglevel';
 import config from './config';
 
-'use strict';
-
-
-var storage = {
-  container: null
+const storage = {
+  container: null,
 };
 
-// set the storage container
-if (config.persist == "local" && window.localStorage) {
-  storage.container = window.localStorage;
-} else if (config.persist == "session" && window.sessionStorage) {
-  storage.container = window.sessionStorage;
-} else if (config.persist == "none") {
-  storage.container = null;
-} else {
-  // Temporary container.
+try {
+  // set the storage container
+  if (config.persist === 'local' && window.localStorage) {
+    storage.container = window.localStorage;
+  } else if (config.persist === 'session' && window.sessionStorage) {
+    storage.container = window.sessionStorage;
+  } else {
+    // "none" or window.localStorage/window.sessionStorage is falsy.
+    storage.container = {};
+  }
+} catch (err) {
+  // Chrome and Edge will throw DOMException when accessing sessionStorage or
+  // localStorage if 3rd party cookie disabled.
+  // https://blog.zok.pw/web/2015/10/21/3rd-party-cookies-in-practice/
+  logger.warn(
+    'Cannot access localStorage/sessionStorage. '
+    + 'This might be due to third-party cookies being disabled.',
+  );
   storage.container = {};
 }
 
 // Pass in accounts from an account manager
-storage.storeAccounts = function (appId, accounts) {
-  if (!storage.container) return;
+storage.storeAccounts = function storeAccounts(appId, accounts) {
+  if (!storage.container) {
+    return;
+  }
 
-  var serviceNames = config.all_services().map(function (service) {
-    return service.id;
-  });
+  const serviceNames = config.all_services().map(service => service.id);
 
-  var key = 'k-' + appId, appData = storage.container[key];
-  if (!appData || appData.length == 0) {
-    appData = {}
+  const key = `k-${appId}`;
+  let appData = storage.container[key];
+  if (!appData || appData.length === 0) {
+    appData = {};
   } else {
     appData = JSON.parse(appData);
   }
 
   // add accounts from the manager for currently visible services.
-  var i, account, array = [];
-  for (i = 0; i < accounts.length; i++) {
-    account = accounts[i];
-    array.push(JSON.stringify(account));
-  }
+  const array = accounts.map(acc => JSON.stringify(acc));
+
   // add accounts already saved for currently invisible services.
-  for (i = 0; i < appData.accounts.length; i++) {
-    account = JSON.parse(appData.accounts[i]);
-    if (serviceNames.indexOf(account.service) == -1) {
-      array.push(appData.accounts[i]);
+  appData.accounts.forEach((acc) => {
+    const account = JSON.parse(acc);
+    if (!serviceNames.includes(account.service)) {
+      array.push(acc);
     }
-  }
+  });
+
   // store the final array
   appData.accounts = array;
   storage.container[key] = JSON.stringify(appData);
@@ -55,13 +60,14 @@ storage.storeAccounts = function (appId, accounts) {
 
 // Return an array of accounts, initialize if necessary
 // the appData is stringified
-storage.loadAccounts = function (appId) {
-  if (!storage.container) return [];
+storage.loadAccounts = function loadAccounts(appId) {
+  if (!storage.container) {
+    return [];
+  }
 
-  var serviceNames = config.all_services().map(function (service) {
-    return service.id;
-  });
-  var key = 'k-' + appId, appData = storage.container[key];
+  const serviceNames = config.all_services().map(service => service.id);
+  const key = `k-${appId}`;
+  let appData = storage.container[key];
   if (!appData || appData.length === 0) {
     appData = {};
     appData.accounts = [];
@@ -71,22 +77,26 @@ storage.loadAccounts = function (appId) {
 
   appData = JSON.parse(appData);
   // accounts also needs to be parsed
-  var i, array = [], accounts = appData.accounts;
-  for (i = 0; i < accounts.length; i++) {
-    var acc = JSON.parse(accounts[i]);
+  const array = [];
+  const { accounts } = appData;
+  for (let i = 0; i < accounts.length; i += 1) {
+    const acc = JSON.parse(accounts[i]);
     if (serviceNames.length === 0 || // Not loaded yet
-      serviceNames.indexOf(acc.service) != -1) {
+      serviceNames.indexOf(acc.service) !== -1) {
       array.push(acc);
     }
   }
   return array;
 };
 
-storage.removeAllAccounts = function (appId) {
-  if (!storage.container) return;
+storage.removeAllAccounts = function removeAllAccounts(appId) {
+  if (!storage.container) {
+    return;
+  }
 
-  var key = 'k-' + appId, appData = storage.container[key];
-  if (!appData || appData.length == 0) {
+  const key = `k-${appId}`;
+  let appData = storage.container[key];
+  if (!appData || appData.length === 0) {
     appData = {};
     appData.accounts = [];
     storage.container[key] = JSON.stringify(appData);
@@ -97,20 +107,20 @@ storage.removeAllAccounts = function (appId) {
   }
 };
 
-storage.storeId = function (pickerId) {
+storage.storeId = function storeId(pickerId) {
   if (!storage.container) {
-    return
-  };
+    return;
+  }
   const key = 'k-explorerId';
   storage.container[key] = pickerId;
-}
+};
 
-storage.loadId = function () {
+storage.loadId = function loadId() {
   if (!storage.container) {
-    return
-  };
+    return null;
+  }
   const key = 'k-explorerId';
   return storage.container[key];
-}
+};
 
 export default storage;