+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Chooser: Introduced the new option, `max_size`, to limit the size of the
+  selected file.
+
 ---
 
 ## [2.5.4] - 2020-10-21