+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+---
+
+## [2.6.0] - 2020-12-14
 ### Added
 - Chooser: Introduced the new option, `max_size`, to limit the size of the
   selected file.
@@ -26,8 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - File Picker's Computer View won't fire success event on cancel or no items
   succeed, to be consistent with other modes.
 - Remove the cancel button in attach mode.
-
----
 
 ## [2.5.4] - 2020-10-21